refactor: 방 목록 조회 페이지 로비 페이지로 이전
