+++ src/App.jsx
@@ -6,7 +6,6 @@ import Logout from "./Logout.jsx";
 import Register from "./Register.jsx";
 import UpdateUser from "./UpdateUser.jsx";
 import DeleteUser from "./DeleteUser.jsx";
-import RoomListPage from "./RoomListPage.jsx";
 import RoomCreatePage from "./RoomCreatePage.jsx";
 import RoomUpdatePage from "./RoomUpdatePage.jsx";
 import ChatRoom from "./ChatRoom.jsx";
@@ -31,7 +30,6 @@ function App() {
                 <Route path="/register" element={<Register />} />
                 <Route path="/update-user" element={<UpdateUser />} />
                 <Route path="/delete-user" element={<DeleteUser />} />
-                <Route path="/rooms" element={<RoomListPage />} />
                 <Route path="/create-room" element={<RoomCreatePage />} />
                 <Route path="/update-room/:roomName" element={<RoomUpdatePage />} />
                 <Route path="/room/:roomName" element={<ChatRoom />} />
